pyproject.toml
########################################
# 1) Системные данные для сборки пакета
########################################
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

########################################
# 2) Метаданные проекта и prod-зависимости
########################################
[project]
name = "jarvis_arduino_bridge"
version = "1.0.0"
description = "Асинхронный мост: **Arduino Mega 2560 ⇄ Serial ⇄ Python ⇄ MQTT ⇄ Home Assistant**"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "Hudrolax", email = "hudro795@gmail.com" }
]

dependencies = [
  "fastapi==0.115.2",
  "uvicorn==0.30.6",
  "PyYAML==6.0.2",
  "pyserial==3.5",
  "pyserial-asyncio==0.6",
  "asyncio-mqtt==0.16.2"
]

########################################
# 3) dev-зависимости
########################################
[project.optional-dependencies]
dev = [
  "pynvim"
]

########################################
# 4) Где искать пакеты (каталог app/)
########################################
[tool.setuptools.packages.find]
where = [""]               # корень пайтон-пакетов
exclude = ["tests*"]

########################################
# 7) Линтеры / форматтеры
########################################
[tool.ruff]
line-length    = 120
target-version = "py312"
select         = ["E", "F", "I"]
extend-select  = ["I"]
src            = ["app"]

[tool.black]
line-length    = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120
lines_after_imports = 2
sections = ["STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
force_sort_within_sections = true
combine_as_imports = false
default_section = "THIRDPARTY"
